fix(llm): multimodal review fixes (module split, dtype cast, config attr, manifest guard)
