+++ packages/llm/torch_to_nnef_llm/models/handlers/__init__.py
@@ -18,8 +18,8 @@ from .idefics3_vl import (
 from .openelm import OpenELMArchitectureHandler
 from .phi import PhiArchitectureHandler
 from .qwen2_5_vl import (
-    Qwen2_5VLArchitectureHandler,
-    Qwen2_5VLVisionEncoderHandler,
+    Qwen25VLArchitectureHandler,
+    Qwen25VLVisionEncoderHandler,
 )
 from .qwen3_vl import (
     Qwen3VLArchitectureHandler,
@@ -51,8 +51,8 @@ __all__ = [
     "MultiModalArchitectureHandler",
     "OpenELMArchitectureHandler",
     "PhiArchitectureHandler",
-    "Qwen2_5VLArchitectureHandler",
-    "Qwen2_5VLVisionEncoderHandler",
+    "Qwen25VLArchitectureHandler",
+    "Qwen25VLVisionEncoderHandler",
     "Qwen35MoeArchitectureHandler",
     "Qwen3VLArchitectureHandler",
     "Qwen3VLVisionEncoderHandler",
